Moved StrykerOptions to api folder and included a way to add custom properties
