+++ src/api/StrykerOptions.ts
@@ -1,6 +1,9 @@
 'use strict';
 
 interface StrykerOptions {
+  // this ensures that custom config for for example 'karma' can be added under the 'karma' key
+  [customConfig: string]: any;
+  
   /**
    * A list of library files.
    */
