+++ src/api/TestRunner.ts
@@ -1,6 +1,9 @@
 import TestResultMap from './test-runner/TestResultMap'
+import StrykerOptions from './StrykerOptions';
 
 interface TestRunner{
+  new(options: StrykerOptions): TestRunner; 
+  
   run(): TestResultMap;
 }
